+++ logging/changelog.md
@@ -1,3 +1,7 @@
+# 0.2.0
+
+- Added the `attachLoggerToContext` middleware and the accompanying `LoggingContext` utility type.
+
 # 0.1.0
 
 - Released package, including `Logger` class.